Use the default value handling of completing-read
On Emacs at least 22.1 and later, completing-read has the 7th argument
DEF which specifies the default value in case the user input is null.
Make use of this argument instead of a home-grown logic to check the
user input later and assign the default value.

This should allow some elisp packages that replace the completion UI,
such as Helm and anything.el, to happily make use of given default
values for their good user experience.